--- destination/entity/category.go
@@ -1,13 +1,0 @@
-package entity
-
-type Category string
-
-const (
-	Analytics      Category = "analytics"
-	Advertising    Category = "advertising"
-	CRM            Category = "crm"
-	EmailMarketing Category = "email-marketing"
-	Livechat       Category = "livechat"
-	Payments       Category = "payments"
-	Surveys        Category = "Surveys"
-)
