--- destination/entity/connectiontype.go
@@ -1,9 +1,0 @@
-package entity
-
-type ConnectionType string
-
-const (
-	EventStream ConnectionType = "event-stream"
-	Storage     ConnectionType = "storage"
-	ReversETL   ConnectionType = "reverse-ETL"
-)
